Finished writing SaveDataManager basic functions
It still needs to be implemented, but I don't have time for that today. It should be able to fully load both the old and new save formats, and it can save to the new save format. It also saves to a tmp file before overwriting the save file, to hopefully avoid corruption.