Minor bug fix
Added a check for Asset file locations by searching for CustomGameMods.scene, which is a requirement.
Fixed a bug where the final dan course in a series would not display on the nameplate when cleared. Possibly fixes (and hopefully not causes) a bug in other areas as well that I didn't look for.