4(ish) items done
Soul Gauge is now properly calculated, displayed, and saved. It is not displayed in the Results screen yet, but it is properly calculated for clear conditions.
Dani Dojo icons now appear on the Song Select menu near songs in the active series.
Requirement bars now properly turn rainbow at the end of the song and/or course for Ok and Bad requirements.
Requirement bars now properly display "Failed" when a requirement has been failed. I still need to play the fail sound though.
Now displays the "Clear Conditions" text in the DaniCourseSelect Scene.